saga and slice moved to personList
